changed homepage to main page
